% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2jupyter.R
\name{rmd2jupyter}
\alias{rmd2jupyter}
\title{Convert Rmd to ipynb (jupyter notebooks)}
\usage{
rmd2jupyter(x)
}
\arguments{
\item{x}{Name of Rmd file}
}
\value{
Saves file with .ipynb extension
}
\description{
Converts Rmd file to a jupyter notebook
}
